feat: implement usage metrics
This is implemented by parsing the mosquitto debug logs
for relevant log entries, and writing them to seperate
log files to be picked up by Log Analytics agent.

This is because there is no plug-in mechanish
that would allow to hook into the mosquitto process.

Log format is defined in https://learn.microsoft.com/en-us/azure/azure-monitor/agents/data-sources-custom-logs